fms-email-campaign-system/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fms-campaigns"
version = "0.1.0"
description = "FMS banner email campaign pipeline (Omnisend + Shopify)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Aaron", email = "info@fabricmegastore.com" }]
license = { text = "Proprietary" }

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.7",
    "sqlmodel>=0.0.22",
    "pillow>=10.3",
    "pytesseract>=0.3.10",
    "rapidfuzz>=3.9",
    "anthropic>=0.34",
    "rich>=13.7",
    "loguru>=0.7",
    "tomli>=2.0; python_version < '3.11'",
    "python-dotenv>=1.0",
    "pyrate-limiter>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.30",
    "pytest-vcr>=1.0.2",
    "vcrpy>=6.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pillow",
]

[project.scripts]
fms-campaigns = "fms_campaigns.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/fms_campaigns"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "live: hits real Omnisend/Shopify (skip in CI)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
